first pass at displaying facepile for likes
both like and repost webmentions are now being displayed as facepiles
above the regular comments.  There's still more to do with the UI, like
better indicating what are likes versus reshares.

The code also needs some cleanup and to be decoupled from Genesis as
much as possible.  It really is already for the most part, I just stuck
this in genesis-indieweb.php for convenience.